CM_Janus_RpcEndpoints: Call to a member function getStreamSubscribes() on a non-object in

+++ library/CM/Janus/RpcEndpoints.php
@@ -82,7 +82,9 @@ class CM_Janus_RpcEndpoints {
 
         $streamRepository = $janus->getStreamRepository();
         $streamChannel = $streamRepository->findStreamChannelByKey($streamChannelKey);
-
+        if (null === $streamChannel) {
+            return;
+        }
         $streamSubscribe = $streamChannel->getStreamSubscribes()->findKey($streamKey);
         if ($streamSubscribe) {
             $streamRepository->removeStream($streamSubscribe);
